+++ lib/config/config.schema.js
@@ -10,7 +10,7 @@ const FieldSchema = object({
   type: string()
     .required('Missing field type')
     .matches(
-      `(${FIELDS.TAGS}|${FIELDS.TEXT}|${FIELDS.TEXTAREA}|${FIELDS.IMAGE}|${FIELDS.FILE}|${FIELDS.SELECT}|${FIELDS.NUMBER}|${FIELDS.JSON}|${FIELDS.BOOLEAN}|${FIELDS.RADIO}|${FIELDS.VIDEO_URL}|${FIELDS.MARKDOWN}|${FIELDS.DATE}|${FIELDS.URL}|${FIELDS.TEL})`,
+      `(${FIELDS.TAGS}|${FIELDS.TEXT}|${FIELDS.TEXTAREA}|${FIELDS.IMAGE}|${FIELDS.FILE}|${FIELDS.SELECT}|${FIELDS.NUMBER}|${FIELDS.JSON}|${FIELDS.BOOLEAN}|${FIELDS.RADIO}|${FIELDS.VIDEO_URL}|${FIELDS.MARKDOWN}|${FIELDS.DATE}|${FIELDS.URL}|${FIELDS.TEL}|${FIELDS.ENTITY_SEARCH})`,
       'Invalid field type'
     ),
   required: boolean().default(false),
@@ -34,6 +34,8 @@ const FieldSchema = object({
   placeholder: string(),
   description: string(),
   hidden: boolean().default(false),
+  entity: string(),
+  entityType: string(),
 })
 
 const ContentTypeSchema = object({
