+++ lib/config/load-config.js
@@ -9,10 +9,6 @@ const ROLES = {
   USER: 'USER'
 }
 
-
-
-
-
 const defaultConfig = {
   title: '', // Empty to throw error if not configured
   description: '', // Empty to throw error if not configured
