+++ lib/constants/entity_types.ts
@@ -1,0 +1,7 @@
+export enum ENTITY_TYPES {
+  USER = 'user',
+  GROUP = 'group',
+  CONTENT = 'content',
+  COMMENT = 'comment',
+  ACTIVITY = 'activity'
+}
