+++ lib/constants/index.ts
@@ -2,3 +2,4 @@ export * from './activity'
 export * from './fields'
 export * from './storages'
 export * from './databases'
+export * from './entity_types'
